fix compilation on gcc 10

+++ src/ge/window/linux/window_xcb.cpp
@@ -108,7 +108,7 @@ namespace ge
                     field = true;
                     ++counter;
                 }
-                state >>= 1;
+                state >>= static_cast<uint16_t>(1);
                 return result;
             };
